--- .changeset/real-coins-accept.md
@@ -1,6 +1,0 @@
----
-"@moderntribe/storebuilder": minor
-"@moderntribe/wme-ui": patch
----
-
-Storebuilder now uses the updated themeing and data model
