--- .changeset/serious-ladybugs-do.md
@@ -1,5 +1,0 @@
----
-"@moderntribe/wme-sitebuilder": minor
----
-
-Refactor how we call wp-cli via the included composer package. Fixes Payment Gateway plugin versions/installation. Fix GitHub actions playwright installation.
