--- .changeset/twenty-needles-sort.md
@@ -1,6 +1,0 @@
----
-"@moderntribe/sitebuilder": major
-"@moderntribe/wme-ui": minor
----
-
-The FTC Wizard now can be completed and will gather all required data from the user
