+++ packages/site-settings/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @moderntribe/sitedetails
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [b01e5af]
+- Updated dependencies [81f8d17]
+  - @moderntribe/wme-ui@3.1.0
+
 ## 1.0.0
 
 ### Major Changes
