+++ packages/site-settings/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@moderntribe/site-settings",
-	"version": "1.0.0",
+	"version": "1.0.1",
 	"description": "Site settings foundation",
 	"source": "src/index.tsx",
 	"main": "dist/umd/index.js",
