+++ packages/sitebuilder/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@moderntribe/sitebuilder",
-	"version": "2.0.1",
+	"version": "3.0.0",
 	"description": "SiteBuilder foundation",
 	"source": "src/index.tsx",
 	"main": "dist/umd/index.js",
