+++ packages/wme-ui/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@moderntribe/wme-ui",
-	"version": "3.0.0",
+	"version": "3.1.0",
 	"private": false,
 	"author": "Moderntribe Incubator Team",
 	"description": "Components and hooks to build the best UX/UI admin wizards",
