+++ plugins/wme-sitebuilder/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@moderntribe/wme-sitebuilder",
-	"version": "2.0.1",
+	"version": "2.1.0",
 	"description": "Easily build and configure a site.",
 	"author": "Modern Tribe Incubator Team",
 	"license": "MIT",
